Close legacy keyring connections on Unix
